Use normal Java conventions for file/package naming. (Class com.foo.bar.Baz should be in com/foo/bar/Baz.java.) To avoid a breaking change, I moved the files to match their existing packages, not the other way around.
Also, moved ContentVerifier to be an inner static class, and added
trailing newlines to appease picky text editors.

Having non-standard file layout, or >1 top level classes per file, makes
incremental builds unreliable since build tools won't know where to look
for a class.
