+++ src/main/java/com/sendgrid/BccSettings.java
@@ -4,9 +4,15 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-@JsonInclude(Include.NON_DEFAULT)
-public class Setting {
+/**
+ * This object allows you to have a blind carbon copy 
+ * automatically sent to the specified email address 
+ * for every email that is sent.
+ */
+@JsonInclude(Include.NON_EMPTY)
+public class BccSettings {
   @JsonProperty("enable") private boolean enable;
+  @JsonProperty("email") private String email;
 
   @JsonProperty("enable")
   public boolean getEnable() {
@@ -17,10 +23,20 @@ public class Setting {
     this.enable = enable;
   }
 
+  @JsonProperty("email")
+  public String getEmail() {
+    return this.email;
+  }
+
+  public void setEmail(String email) {
+    this.email = email;
+  }
+
   @Override
   public int hashCode() {
     final int prime = 31;
     int result = 1;
+    result = prime * result + ((email == null) ? 0 : email.hashCode());
     result = prime * result + (enable ? 1231 : 1237);
     return result;
   }
@@ -33,7 +49,12 @@ public class Setting {
       return false;
     if (getClass() != obj.getClass())
       return false;
-    Setting other = (Setting) obj;
+    BccSettings other = (BccSettings) obj;
+    if (email == null) {
+      if (other.email != null)
+        return false;
+    } else if (!email.equals(other.email))
+      return false;
     if (enable != other.enable)
       return false;
     return true;
