+++ src/main/java/com/sendgrid/ClickTrackingSetting.java
@@ -5,39 +5,38 @@ import com.fasterxml.jackson.annotation.JsonInclude.Include;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 /**
- * This object allows you to have a blind carbon copy 
- * automatically sent to the specified email address 
- * for every email that is sent.
+ * Settings to determine how you would like to track the 
+ * metrics of how your recipients interact with your email.
  */
 @JsonInclude(Include.NON_EMPTY)
-public class BccSettings {
+public class ClickTrackingSetting {
   @JsonProperty("enable") private boolean enable;
-  @JsonProperty("email") private String email;
-
+  @JsonProperty("enable_text") private boolean enableText;
+ 
   @JsonProperty("enable")
   public boolean getEnable() {
     return enable;
   }
-
+  
   public void setEnable(boolean enable) {
     this.enable = enable;
   }
-
-  @JsonProperty("email")
-  public String getEmail() {
-    return this.email;
-  }
-
-  public void setEmail(String email) {
-    this.email = email;
+  
+  @JsonProperty("enable_text")
+  public boolean getEnableText() {
+    return enableText;
+  }  
+  
+  public void setEnableText(boolean enableText) {
+    this.enableText = enableText;
   }
 
   @Override
   public int hashCode() {
     final int prime = 31;
     int result = 1;
-    result = prime * result + ((email == null) ? 0 : email.hashCode());
     result = prime * result + (enable ? 1231 : 1237);
+    result = prime * result + (enableText ? 1231 : 1237);
     return result;
   }
 
@@ -49,13 +48,10 @@ public class BccSettings {
       return false;
     if (getClass() != obj.getClass())
       return false;
-    BccSettings other = (BccSettings) obj;
-    if (email == null) {
-      if (other.email != null)
-        return false;
-    } else if (!email.equals(other.email))
-      return false;
+    ClickTrackingSetting other = (ClickTrackingSetting) obj;
     if (enable != other.enable)
+      return false;
+    if (enableText != other.enableText)
       return false;
     return true;
   }
