+++ src/main/java/com/sendgrid/Content.java
@@ -4,11 +4,8 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-import java.util.Arrays;
 import java.util.Collections;
-import java.util.HashSet;
 import java.util.List;
-import java.util.Set;
 import java.util.regex.Pattern;
 import java.lang.IllegalArgumentException;
 
@@ -78,18 +75,19 @@ public class Content {
       return false;
     return true;
   }
-}
 
-class ContentVerifier {
-  private static final List<Pattern> FORBIDDEN_PATTERNS = Collections.singletonList(
-          Pattern.compile(".*SG\\.[a-zA-Z0-9(-|_)]*\\.[a-zA-Z0-9(-|_)]*.*")
-  );
+  static class ContentVerifier {
+    private static final List<Pattern> FORBIDDEN_PATTERNS = Collections.singletonList(
+            Pattern.compile(".*SG\\.[a-zA-Z0-9(-|_)]*\\.[a-zA-Z0-9(-|_)]*.*")
+    );
 
-  static void verifyContent(String content) {
-    for (Pattern pattern: FORBIDDEN_PATTERNS) {
-      if (pattern.matcher(content).matches()) {
-        throw new IllegalArgumentException("Found a Forbidden Pattern in the content of the email");
+    static void verifyContent(String content) {
+      for (Pattern pattern: FORBIDDEN_PATTERNS) {
+        if (pattern.matcher(content).matches()) {
+          throw new IllegalArgumentException("Found a Forbidden Pattern in the content of the email");
+        }
       }
     }
   }
 }
+
