+++ src/main/java/com/sendgrid/GoogleAnalyticsSetting.java
