+++ src/main/java/com/sendgrid/OpenTrackingSetting.java
@@ -5,14 +5,16 @@ import com.fasterxml.jackson.annotation.JsonInclude.Include;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 /**
- * Settings to determine how you would like to track the 
- * metrics of how your recipients interact with your email.
+ * An open tracking settings object. This allows you to track 
+ * whether the email was opened or not, but including a single 
+ * pixel image in the body of the content. When the pixel is 
+ * loaded, we can log that the email was opened.
  */
 @JsonInclude(Include.NON_EMPTY)
-public class ClickTrackingSetting {
+public class OpenTrackingSetting {
   @JsonProperty("enable") private boolean enable;
-  @JsonProperty("enable_text") private boolean enableText;
- 
+  @JsonProperty("substitution_tag") private String substitutionTag;
+  
   @JsonProperty("enable")
   public boolean getEnable() {
     return enable;
@@ -22,13 +24,13 @@ public class ClickTrackingSetting {
     this.enable = enable;
   }
   
-  @JsonProperty("enable_text")
-  public boolean getEnableText() {
-    return enableText;
-  }  
+  @JsonProperty("substitution_tag")
+  public String getSubstitutionTag() {
+    return substitutionTag;
+  }
   
-  public void setEnableText(boolean enableText) {
-    this.enableText = enableText;
+  public void setSubstitutionTag(String substitutionTag) {
+    this.substitutionTag = substitutionTag;
   }
 
   @Override
@@ -36,7 +38,7 @@ public class ClickTrackingSetting {
     final int prime = 31;
     int result = 1;
     result = prime * result + (enable ? 1231 : 1237);
-    result = prime * result + (enableText ? 1231 : 1237);
+    result = prime * result + ((substitutionTag == null) ? 0 : substitutionTag.hashCode());
     return result;
   }
 
@@ -48,11 +50,15 @@ public class ClickTrackingSetting {
       return false;
     if (getClass() != obj.getClass())
       return false;
-    ClickTrackingSetting other = (ClickTrackingSetting) obj;
+    OpenTrackingSetting other = (OpenTrackingSetting) obj;
     if (enable != other.enable)
       return false;
-    if (enableText != other.enableText)
+    if (substitutionTag == null) {
+      if (other.substitutionTag != null)
+        return false;
+    } else if (!substitutionTag.equals(other.substitutionTag))
       return false;
     return true;
   }
+  
 }
