+++ src/main/java/com/sendgrid/Setting.java
@@ -1,0 +1,41 @@
+package com.sendgrid;
+
+import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.annotation.JsonInclude.Include;
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+@JsonInclude(Include.NON_DEFAULT)
+public class Setting {
+  @JsonProperty("enable") private boolean enable;
+
+  @JsonProperty("enable")
+  public boolean getEnable() {
+    return enable;
+  }
+
+  public void setEnable(boolean enable) {
+    this.enable = enable;
+  }
+
+  @Override
+  public int hashCode() {
+    final int prime = 31;
+    int result = 1;
+    result = prime * result + (enable ? 1231 : 1237);
+    return result;
+  }
+
+  @Override
+  public boolean equals(Object obj) {
+    if (this == obj)
+      return true;
+    if (obj == null)
+      return false;
+    if (getClass() != obj.getClass())
+      return false;
+    Setting other = (Setting) obj;
+    if (enable != other.enable)
+      return false;
+    return true;
+  }
+}
