+++ src/main/java/com/sendgrid/SpamCheckSetting.java
