+++ src/main/java/com/sendgrid/SubscriptionTrackingSetting.java
@@ -5,20 +5,24 @@ import com.fasterxml.jackson.annotation.JsonInclude.Include;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 /**
- * An object representing the default footer 
- * that you would like included on every email.
+ * A subscription tracking setting object. Subscription tracking 
+ * allows you to insert a subscription management link at the 
+ * bottom of the text and html bodies of your email. If you 
+ * would like to specify the location of the link within your 
+ * email, you may use the substitution_tag.
  */
 @JsonInclude(Include.NON_EMPTY)
-public class FooterSetting {
+public class SubscriptionTrackingSetting {
   @JsonProperty("enable") private boolean enable;
   @JsonProperty("text") private String text;
   @JsonProperty("html") private String html;
+  @JsonProperty("substitution_tag") private String substitutionTag;
   
   @JsonProperty("enable")
   public boolean getEnable() {
     return enable;
   }
-
+ 
   public void setEnable(boolean enable) {
     this.enable = enable;
   }
@@ -35,10 +39,19 @@ public class FooterSetting {
   @JsonProperty("html")
   public String getHtml() {
     return html;
-  }
+  } 
 
   public void setHtml(String html) {
     this.html = html;
+  }
+  
+  @JsonProperty("substitution_tag")
+  public String getSubstitutionTag() {
+    return substitutionTag;
+  }  
+
+  public void setSubstitutionTag(String substitutionTag) {
+    this.substitutionTag = substitutionTag;
   }
 
   @Override
@@ -47,6 +60,7 @@ public class FooterSetting {
     int result = 1;
     result = prime * result + (enable ? 1231 : 1237);
     result = prime * result + ((html == null) ? 0 : html.hashCode());
+    result = prime * result + ((substitutionTag == null) ? 0 : substitutionTag.hashCode());
     result = prime * result + ((text == null) ? 0 : text.hashCode());
     return result;
   }
@@ -59,13 +73,18 @@ public class FooterSetting {
       return false;
     if (getClass() != obj.getClass())
       return false;
-    FooterSetting other = (FooterSetting) obj;
+    SubscriptionTrackingSetting other = (SubscriptionTrackingSetting) obj;
     if (enable != other.enable)
       return false;
     if (html == null) {
       if (other.html != null)
         return false;
     } else if (!html.equals(other.html))
+      return false;
+    if (substitutionTag == null) {
+      if (other.substitutionTag != null)
+        return false;
+    } else if (!substitutionTag.equals(other.substitutionTag))
       return false;
     if (text == null) {
       if (other.text != null)
