+++ src/test/java/com/sendgrid/MailTest.java
@@ -3,7 +3,6 @@ package com.sendgrid;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
 import org.junit.Assert;
-import org.junit.Before;
 import org.junit.Test;
 
 import java.io.IOException;
